fix(game): impedindo a renderização de colisões desativadas

+++ src/game/src/HitboxLoader.cpp
@@ -59,7 +59,8 @@ Entity HitboxLoader::createHitboxFromNode(const DataNode& node, Entity parent, b
         Rectangle rect = DataUtils::parseRect(node);
         LOG_DEBUG("HitboxLoader: parsed rect size=({},{}) offset=({},{}) debug.color=({},{},{},{})",
             rect.size.width, rect.size.height, offset.x, offset.y,
-            debug.color.r, debug.color.g, debug.color.b, debug.color.a);
+            static_cast<int>(debug.color.r), static_cast<int>(debug.color.g),
+            static_cast<int>(debug.color.b), static_cast<int>(debug.color.a));
         return this->factory.createHitboxChild(EntityFactory::HitboxChildParams{
             parent, offset, damage, facingLeft, debug}, rect);
     }
