+++ src/game/src/PushboxLoader.cpp
@@ -61,7 +61,8 @@ Entity PushboxLoader::createPushboxFromNode(const DataNode& node, Entity parent,
         Rectangle rect = DataUtils::parseRect(node);
         LOG_DEBUG("PushboxLoader: parsed rect size=({},{}) offset=({},{}) debug.color=({},{},{},{})",
             rect.size.width, rect.size.height, offset.x, offset.y,
-            debug.color.r, debug.color.g, debug.color.b, debug.color.a);
+           static_cast<int>(debug.color.r), static_cast<int>(debug.color.g),
+           static_cast<int>(debug.color.b), static_cast<int>(debug.color.a));
         return this->factory.createPushboxChild(EntityFactory::PushboxChildParams{
             parent, offset, pushType, mass, pushResistance, facingLeft, debug}, rect);
     }
